add new tst

+++ examples/tst-tw-timer/heapopr.cpp
@@ -301,7 +301,7 @@ timer_heap::~timer_heap() {
 
 
 void printVec(vector<int> nums) {
-    for (int i = 0; i < nums.size(); ++i)
+    for (size_t i = 0; i < nums.size(); ++i)
         cout << nums[i] << " ";
     cout << endl;
 }
